ml/requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
catboost>=1.2.0
xgboost>=2.0.0
pymysql>=1.1.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
